requirements.txt
fastapi[standard]
psycopg2-binary
sqlmodel
pyjwt
pydantic[email]
pytest
bcrypt